+++ packages/xyd-cli/tsup.config.ts
@@ -24,6 +24,7 @@ import themeGustoPkg from '../xyd-theme-gusto/package.json';
 import themePoetryPkg from '../xyd-theme-poetry/package.json';
 import uiPkg from '../xyd-ui/package.json';
 import uniformPkg from '../xyd-uniform/package.json';
+import sourcesPk from '../xyd-sources/package.json';
 
 const entry = ["index.ts"];
 
@@ -49,6 +50,7 @@ const deps = {
         ...themePoetryPkg.dependencies,
         ...uiPkg.dependencies,
         ...uniformPkg.dependencies,
+        ...sourcesPk.dependencies,
     },
     dev: {
         ...cliPkg.devDependencies,
@@ -68,6 +70,7 @@ const deps = {
         ...themePoetryPkg.devDependencies,
         ...uiPkg.devDependencies,
         ...uniformPkg.devDependencies,
+        ...sourcesPk.devDependencies,
     }
 }
 
@@ -150,7 +153,7 @@ export default defineConfig([
             execSync(`cp -r ../xyd-plugin-zero/src/pages/* ${pluginPagesDir}`);
 
             // TODO: DEV AND PROD MOD
-            if (process.env.XYD_DEV_MODE) {
+            if (process.env.XYD_DEV_MODE && !process.env.XYD_DEV_CLI_NOINSTALL) {
                 execSync(`cd ${cliDir} && pnpm i && pnpm link --global`);
             } else {
                 // for prod npm?