+++ packages/xyd-content/packages/md/index.ts
@@ -1,7 +1,7 @@
-import {RemarkMdxTocOptions} from "./plugins/md-toc";
+import {RemarkMdxTocOptions} from "./plugins/mdToc";
 import {defaultPlugins} from "./plugins"
 
-export {RemarkMdxTocOptions} from "./plugins/md-toc";
+export {RemarkMdxTocOptions} from "./plugins/mdToc";
 
 export function mdOptions(toc: RemarkMdxTocOptions) {
     return {