+++ packages/xyd-sources/example/package-b/src/index.ts
@@ -1,0 +1,8 @@
+export * from "./billing"
+
+/**
+ * Returns a message builder message.
+ */
+export function messageBuilder(): string {
+    return 'Hello world!';
+}