+++ packages/xyd-theme-poetry/src/components/Layouts/index.ts
@@ -1,5 +1,0 @@
-export {
-    Layout
-} from "./Layout";
-
-export type {LayoutProps} from "./Layout";