+++ packages/fable-wiki/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @xyd-js/fable-wiki
 
+## 0.1.0-xyd.10
+
+### Patch Changes
+
+- Updated dependencies [b9651b3]
+  - @xyd-js/framework@0.1.0-xyd.27
+
+## 0.1.0-xyd.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.7
+  - @xyd-js/framework@0.1.0-xyd.26
+
+## 0.1.0-xyd.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.6
+  - @xyd-js/framework@0.1.0-xyd.25
+
+## 0.1.0-xyd.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.5
+  - @xyd-js/framework@0.1.0-xyd.24
+
 ## 0.1.0-xyd.6
 
 ### Patch Changes