+++ packages/xyd-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @xyd-js/core
 
+## 0.1.0-xyd.9
+
+### Patch Changes
+
+- refactor, tweaks and ui changes
+
+## 0.1.0-xyd.8
+
+### Patch Changes
+
+- ui tweaks
+
+## 0.1.0-xyd.7
+
+### Patch Changes
+
+- ui tweaks, and some code refactor
+
 ## 0.1.0-xyd.6
 
 ### Patch Changes