+++ packages/xyd-framework/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @xyd-js/framework
 
+## 0.1.0-xyd.27
+
+### Patch Changes
+
+- b9651b3: version
+
+## 0.1.0-xyd.26
+
+### Patch Changes
+
+- refactor, tweaks and ui changes
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.7
+  - @xyd-js/content@0.1.0-xyd.10
+  - @xyd-js/core@0.1.0-xyd.9
+  - @xyd-js/ui@0.1.0-xyd.12
+
+## 0.1.0-xyd.25
+
+### Patch Changes
+
+- ui tweaks
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.6
+  - @xyd-js/content@0.1.0-xyd.9
+  - @xyd-js/core@0.1.0-xyd.8
+  - @xyd-js/ui@0.1.0-xyd.11
+
+## 0.1.0-xyd.24
+
+### Patch Changes
+
+- ui tweaks, and some code refactor
+- Updated dependencies
+  - @xyd-js/components@0.1.0-xyd.5
+  - @xyd-js/content@0.1.0-xyd.8
+  - @xyd-js/core@0.1.0-xyd.7
+  - @xyd-js/ui@0.1.0-xyd.10
+
 ## 0.1.0-xyd.23
 
 ### Patch Changes