+++ packages/xyd-openapi/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @xyd-js/openapi
 
+## 0.1.0-xyd.7
+
+### Patch Changes
+
+- refactor, tweaks and ui changes
+- Updated dependencies
+  - @xyd-js/uniform@0.1.0-xyd.11
+
+## 0.1.0-xyd.6
+
+### Patch Changes
+
+- @xyd-js/uniform@0.1.0-xyd.10
+
+## 0.1.0-xyd.5
+
+### Patch Changes
+
+- @xyd-js/uniform@0.1.0-xyd.9
+
 ## 0.1.0-xyd.4
 
 ### Patch Changes