+++ packages/xyd-theme-poetry/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @xyd-js/theme-poetry
 
+## 0.1.0-xyd.23
+
+### Patch Changes
+
+- b9651b3: version
+- Updated dependencies [b9651b3]
+  - @xyd-js/framework@0.1.0-xyd.27
+
+## 0.1.0-xyd.22
+
+### Patch Changes
+
+- refactor, tweaks and ui changes
+- Updated dependencies
+  - @xyd-js/atlas@0.1.0-xyd.12
+  - @xyd-js/components@0.1.0-xyd.7
+  - @xyd-js/framework@0.1.0-xyd.26
+
+## 0.1.0-xyd.21
+
+### Patch Changes
+
+- ui tweaks
+- Updated dependencies
+  - @xyd-js/atlas@0.1.0-xyd.11
+  - @xyd-js/components@0.1.0-xyd.6
+  - @xyd-js/framework@0.1.0-xyd.25
+
+## 0.1.0-xyd.20
+
+### Patch Changes
+
+- ui tweaks, and some code refactor
+- Updated dependencies
+  - @xyd-js/atlas@0.1.0-xyd.10
+  - @xyd-js/components@0.1.0-xyd.5
+  - @xyd-js/framework@0.1.0-xyd.24
+
+## 0.1.0-xyd.19
+
+### Patch Changes
+
+- update @xyd-js/react-router-dev
+
 ## 0.1.0-xyd.18
 
 ### Patch Changes